Added Color Blob Detection
Our color blob detection works by first defining 2 zones on the camera frame (where the spike marks are). Then, the pipeline takes the average color in the zones. Next, we compare these colors to red/blue, depending on the alliance you are on. The zone color that is closest to red/blue is where the team prop is. But there is no zone covering the right spike mark. So we see if the zones distances from red/blue are past a certain value that decides if the prop is in the right spike mark. To learn more, watch our explanation video. For the code, look at our GitHub.